BigAnimal and driver3 updates (#12)
* Ignore non-zero return code from grep -c

In the logic to wait until the BigAnimal cluster is ready, we use `grep
-c` to determine when the cluster is ready.  This returns a non-zero
code if nothing is found, which will make `bash -e` exit the script at
that point.

* Update playbooks to use driver3

driver3 uses multi-process event-driven models to drive the workload,
significantly reducing hardware resources needed to drive a test.  This
can be significant with 10,000+ warehouses.  This is a 2-tier
client-server configuration that removes the need for the "client" tier.

Use r5.8xlarge database systems to test.

Use c5.9xlarge systems to drive the databases, matching the processor
count (32) of the r5.8xlarge class systems.  This is intended to
minimize loading time, which may be the most significant time sink in
testing.

Set the sleep time to 300ms (reduced from default of 1) to decrease ramp
up time, but still give enough time to warm the database before the
measurement interval.

* Add aws-dbt2-biganimal/prepare directory

This should have been a part of 7980961b798e3974dc1ee03f9ac2fc7892c04b65
but most have been dropped by accident.