Correctly fix dbt2 db connection controls (#23)
* Revert "Use dbt2_connections to control -t flag"

This reverts commit ddc27baa9506594aac89cc55355076513bae5e82.

* Correctly control the number of dbt2 database connections

The event driven driver for dbt2 opens a database connection per driver
process spawned.  The driver spawns 1 process per processor on the
system, which are pinned per process.

The event driven driver can be told how many processes to spawn per
processor with the -F flag to the run-workload script.

Calculate the -F value based on how many connections are set to be
opened by the buildbot worker.  Note this will be rounded down, if not
set to be a multiple of the number of processors on the driver system.

* Target 45 minute rampup

A 10,000 warehouse database should warm up around 30 minutes.  Target 45
minutes to be sure the database is warm before the measurement interval.